feat: add BvhManagerPlugin, split bvh_gizmos from bvh_player
