--- src/bvh.rs
@@ -1,2 +1,0 @@
-pub mod bvh_library;
-pub mod bvh_player;
